+++ src/core/action.ts
@@ -46,7 +46,7 @@ function startAction(
 ): IActionRunInfo {
     const notifySpy = isSpyEnabled() && !!actionName
     let startTime: number = 0
-    if (notifySpy) {
+    if (notifySpy && process.env.NODE_ENV !== "production") {
         startTime = Date.now()
         const l = (args && args.length) || 0
         const flattendArgs = new Array(l)
@@ -73,7 +73,8 @@ function endAction(runInfo: IActionRunInfo) {
     allowStateChangesEnd(runInfo.prevAllowStateChanges)
     endBatch()
     untrackedEnd(runInfo.prevDerivation)
-    if (runInfo.notifySpy) spyReportEnd({ time: Date.now() - runInfo.startTime })
+    if (runInfo.notifySpy && process.env.NODE_ENV !== "production")
+        spyReportEnd({ time: Date.now() - runInfo.startTime })
 }
 
 export function allowStateChanges<T>(allowStateChanges: boolean, func: () => T): T {