+++ src/core/reaction.ts
@@ -90,7 +90,11 @@ export class Reaction implements IDerivation, IReactionPublic {
 
                 try {
                     this.onInvalidate()
-                    if (this._isTrackPending && isSpyEnabled()) {
+                    if (
+                        this._isTrackPending &&
+                        isSpyEnabled() &&
+                        process.env.NODE_ENV !== "production"
+                    ) {
                         // onInvalidate didn't trigger track right away..
                         spyReport({
                             name: this.name,
@@ -109,7 +113,7 @@ export class Reaction implements IDerivation, IReactionPublic {
         startBatch()
         const notify = isSpyEnabled()
         let startTime
-        if (notify) {
+        if (notify && process.env.NODE_ENV !== "production") {
             startTime = Date.now()
             spyReportStart({
                 name: this.name,
@@ -125,7 +129,7 @@ export class Reaction implements IDerivation, IReactionPublic {
             clearObserving(this)
         }
         if (isCaughtException(result)) this.reportExceptionInDerivation(result.cause)
-        if (notify) {
+        if (notify && process.env.NODE_ENV !== "production") {
             spyReportEnd({
                 time: Date.now() - startTime
             })